файл переименован, компонент разбит на более мелкие компоненты
